Implemented a basic attack for dash when dashing right
Unfortunately, the dash is still a work in progress. Wall clipping is still possible, but minimized due to raycasts pointing left and right, which measure the distance between player and platforms. I also added a basic attack, so that anything behind the player is damaged while dashing. The way the dash currently functions is that it moves fast enough to pass through various game objects. Raycasts can be used to prevent the player from clipping through walls. The main things that remain to be implemented are i-frames for the player passing through enemies and cooldown. Everything else simply requires polish.